+++ CHANGELOG.md
@@ -5,12 +5,13 @@
 - Added Hotspot Pointer with customizable priority.
   - Use at your own risk! Off by default
 - Added hotspot sharing
+- Added party commands
 
 ### Changes: 
 - Made the report/delete button in rfupf last a little bit so it is easier to click
 
 ### Fixes:
-- Made party tracking also work with /stream command
+- Made party tracking also work with /stream command and upon inviting someone
 
 ### Back-end:
 - Added line rendering